Reformatting & rewording/rephrasing
Added -v option--displays output deemed excessive/verbose/for debugging

Added label for help menu in help menu

Added word count info to dictionary download prompt

Added stdwrn & stdver (FDs 3 & 4)

Moved errors to stderr, warnings and misc. info to stdwrn.

Implemented searching of dictionaries in /usr/share/dict/ before falling back to MIT dictionary

+++ hardening/passwd-gen.sh
@@ -3,26 +3,50 @@
 # Environment Setup
 #
 # File name/location of dictionary file
-dict_location='en_US-dict.txt'
+dict_location="$(find /usr/share/dict -type f,l -readable | head -n1)"
 dict_url='https://www.mit.edu/~ecprice/wordlist.10000'
+#
+# Duplicate stderr from FD 2 to FD 3, FD 3 will be used as "stdver" (standard verbose)
+exec 3>&2
+#
+# Duplicate stderr from FD 2 to FD 4, FD 4 will be used as "stdwrn" (standard warning)
+exec 4>&2
 #
 # Helper function to pull a random word
 word_pull() {
 	# Pull a word within length constraints
 	local word kill
-	while [[ ${#word} -lt "$min" || ${#word} -gt "$max" ]]; do
-		if [[ "$kill" -gt 100 ]]; then
-			# Error out if failed to find a word matching set constraints for more than 100 cycles
-			echo 'E: Script took too long to find a word within constraints, stopping to avoid system stress...' >&2
-			return 3
+	while [[
+			${#word} -lt "$min" ||
+				${#word} -gt "$max"
+		]]; do
+		#
+		# Verbose output: print scrapped words
+		if [[
+				-n "$verbose" &&
+					-n "$word"
+			]]; then
+			echo "i: Hit an unfit word '$word'" >&3
 		fi
-		word=$(
-			if ! shuf -n1 "$dict_location"; then
-				# Error out if failed to fetch a word
-				echo 'E: Failed to fetch word from dictionary.' >&2
-				return 2
-			fi
-		)
+		#
+		# Error out if failed to find a word matching set constraints for more than 250 cycles
+		if [[ "$kill" -gt 250 ]]; then
+			echo 'E: Script took too long to find a word within constraints, stopping to avoid system stress...' >&4
+			return 4
+		fi
+		#
+		# Shuffle the dictionary and pull 1 word
+		if ! word=$(
+				if ! shuf -n1 "$dict_location"; then
+					# Error out if failed to fetch a word
+					echo 'E: Failed to fetch word from the dictionary.' >&4
+					return 2
+				fi
+			); then
+			return 2
+		fi
+		#
+		# Increase cycle count
 		(( kill++ ))
 	done
 	#
@@ -36,7 +60,7 @@ word_pull() {
 }
 #
 # Parse CLI arguments
-while getopts 's:m:M:c:a:p:h' flag; do
+while getopts 's:m:M:c:a:p:hv' flag; do
 	case "${flag}" in
 		s)
 			separator="${OPTARG}"
@@ -56,8 +80,14 @@ while getopts 's:m:M:c:a:p:h' flag; do
 		p)
 			pattern="${OPTARG}"
 		;;
-		h)
-			cat <<-'EOF'
+		v)
+			verbose='y'
+		;;
+		h|*)
+			cat >&4 <<-'EOF'
+				You're looking at: The Help Menu
+				Either you used a non-existent option or asked to be directed here by way of -h.
+
 				About:
 				    A very customizable and automatible password generator
 				
@@ -83,6 +113,9 @@ while getopts 's:m:M:c:a:p:h' flag; do
 				    -p: Requires an argument (Any string is technically allowed but characters other than w/W/n/N/s/S are ignored in operation)
 				        Determines the arrangement of the password
 
+				    -v: Requires NO argument
+				        Displays verbose output
+
 				    -h: Requires NO argument
 				        Displays this information screen
 				
@@ -102,9 +135,6 @@ while getopts 's:m:M:c:a:p:h' flag; do
 			EOF
 			exit 0
 		;;
-		*)
-			echo 'W: An invalid option was ignored.' >&2
-		;;
 	esac
 done
 
@@ -116,59 +146,79 @@ done
 # Generation Tuning
 #
 # Prompts for tuning
-if [[ -z "$separator" ]]; then
-	read -rp 'Enter a separator (Default is -): ' separator
+if [[
+		-z "$separator" &&
+			-t 0
+	]]; then
+	read -rp 'Enter a separator (Default is -): ' separator >&4
 fi
-if [[ -z "$min" ]]; then
-	read -rp 'Minimum word length (Default is 4): ' min
+if [[
+		-z "$min" &&
+			-t 0
+	]]; then
+	read -rp 'Minimum word length (Default is 4): ' min >&4
 fi
-if [[ -z "$max" ]]; then
-	read -rp 'Max word length (Default is 8): ' max
+if [[
+		-z "$max" &&
+			-t 0
+	]]; then
+	read -rp 'Max word length (Default is 8): ' max >&4
 fi
-if [[ -z "$capitals" ]]; then
-	read -rp 'Capitalize first letters? [y/n] (Default is yes): ' capitals
+if [[
+		-z "$capitals" &&
+			-t 0
+	]]; then
+	read -rp 'Capitalize first letters? [Y/n]: ' capitals >&4
 fi
-if [[ -z "$password_amount" ]]; then
-	read -rp 'Amount of passwords (Default is 1): ' password_amount
+if [[
+		-z "$password_amount" &&
+			-t 0
+	]]; then
+	read -rp 'Amount of passwords (Default is 1): ' password_amount >&4
 fi
-if [[ -z "$pattern" ]]; then
-	cat <<-'EOF'
+if [[
+		-z "$pattern" &&
+			-t 0
+	]]; then
+	cat >&4 <<-'EOF'
 		w = Random word
 		n = Random number
 		s = Provided separator
 	EOF
-	read -rp 'Enter your generation pattern (Default is wnswnswn): ' pattern
+	read -rp 'Enter your generation pattern (Default is wnswnswn): ' pattern >&4
 fi
 #
 # Validate inputs
 if [[ -z "$separator" ]]; then
-	echo 'i: No separator given; defaulting to a hyphen (-)' >&2
+	echo 'i: No separator given; defaulting to a hyphen (-)' >&4
 	separator='-'
 fi
 if [[ ! "$min" =~ ^[0-9]+$ ]]; then
-	echo 'i: Invalid/missing minimum length; defaulting to 4.' >&2
+	echo 'i: Invalid/missing minimum length; defaulting to 4.' >&4
 	min=4
 fi
 if [[ ! "$max" =~ ^[0-9]+$ ]]; then
-	echo 'i: Invalid/missing maximum length; defaulting to 8.' >&2
+	echo 'i: Invalid/missing maximum length; defaulting to 8.' >&4
 	max=8
 fi
 if (( "$min" > "$max" )); then
-	echo "W: Minimum ($min) is greater than maximum ($max); setting min/max to default values of 4 & 8 respectively and proceeding..." >&2
-	min=4
-	max=8
+	cat >&4 <<-EOF
+		W: Minimum ($min) is greater than maximum ($max) is an unfufilable condition;
+		i: Swapping the values of min/max from $min/$max to $max/$min to fix contradiction and proceeding...
+	EOF
+	read -r max min <<< "$(echo $min $max)"
 fi
 capitals="${capitals:0:1}"
 if [[ ! "$capitals" =~ ^[yYnN]$ ]]; then
-	echo 'i: Invalid/missing response; defaulting to [Y]es.' >&2
+	echo 'i: Invalid/missing response; defaulting to [Y]es.' >&4
 	capitals='Y'
 fi
 if [[ -z "$pattern" ]]; then
-	echo 'i: Invalid/missing response; defaulting to wnswnswn.' >&2
+	echo 'i: Invalid/missing response; defaulting to wnswnswn.' >&4
 	pattern='wnswnswn'
 fi
 if [[ "$password_amount" -lt 1 ]]; then
-	echo 'i: Invalid/missing response; defaulting to 1.' >&2
+	echo 'i: Invalid/missing response; defaulting to 1.' >&4
 	password_amount=1
 fi
 
@@ -180,18 +230,37 @@ fi
 # Password generation
 #
 # Download dictionary if not already present
-if [[ ! -f "$dict_location" ]]; then
-	echo "W: Dictionary does not exist. Downloading a dictionary from '$dict_url' (aprox. ~76kb of data)..." >&2
-	if ! curl -s "$dict_url" --connect-timeout 5 > "$dict_location"; then
-		if [[ -f "$dict_location" ]]; then
-			rm "$dict_location"
+if [[ ! -f "$dict_location" && ! -f 'en_US-dict.txt' ]]; then
+	echo "W: A pre-existing dictionary couldn't be located in '$dict_location'."
+	dict_location='en_US-dict.txt'
+	if [[ -t 0 ]]; then
+		read -rp "Download a dictionary from '$dict_url' to '$(pwd)/$dict_location? (aprox. ~76kb of characters, 10k words) [y/N]: '" download >&4
+	else
+		cat >&4 <<-'EOF'
+			E: As this is non-interactive,
+			    a prompt won't be shown for downloading an external dictionary,
+			    the dictionary won't be downloaded
+			    and the script will now close.
+		EOF
+		exit 4
+	fi
+	if [[ "$download" =~ ^[yY] ]]; then
+		echo 'i: Downloading...' >&4
+		if ! curl -s "$dict_url" --connect-timeout 5 > "$dict_location"; then
+			cat >&4 <<-'EOF'
+				E: Failed to download dictionary;
+				    deleting possible remnant file(s) and quitting...
+			EOF
+			if [[ -f "$dict_location" ]]; then
+				rm "$dict_location"
+			fi
+			exit 1
 		fi
-		exit 1
 	fi
 fi
 #
 # Generate password(s)
-echo 'Generated password(s):' >&2
+echo 'Generated password(s):' >&4
 for (( x=0; x < password_amount; x++ )); do
 	result=
 	for (( y=0; y < ${#pattern}; y++ )); do
@@ -199,7 +268,8 @@ for (( x=0; x < password_amount; x++ )); do
 		case "$char" in
 			w|W)
 				# Parse w/W into a random word
-				word=$(word_pull) || exit "$?"
+				word=$(word_pull) ||
+					exit "$?"
 				result+="$word"
 			;;
 			n|N)
@@ -212,7 +282,7 @@ for (( x=0; x < password_amount; x++ )); do
 			;;
 			*)
 				# More input validation
-				echo "W: Unrecognized character '$char' in pattern at line 1, column $y. Ignoring..." >&2
+				echo "W: Unrecognized character '$char' in pattern at line 1, column $y. Ignoring..." >&4
 			;;
 		esac
 	done
